+++ src/shinestacker/app/project.py
@@ -53,10 +53,14 @@ def main():
         prog=f'{constants.APP_STRING.lower()}-project',
         description='Manage and run focus stack jobs.',
         epilog=f'This app is part of the {constants.APP_STRING} package.')
-    setup_filename_argument(parser, use_const=True)
+    setup_filename_argument(parser)
     add_project_arguments(parser)
     args = vars(parser.parse_args(filtered_args))
     filename = process_filename_argument(args, positional_filename)
+    path = args['path']
+    if filename and path:
+        print("can't specify both arguments --filename and --path", file=sys.stderr)
+        sys.exit(1)
     app = make_app(Application)
     window = ProjectApp()
     if args['expert']:
@@ -65,6 +69,8 @@ def main():
     window.show()
     if filename:
         QTimer.singleShot(100, lambda: window.open_project(filename))
+    elif path:
+        QTimer.singleShot(100, lambda: window.new_project(path))
     elif args['new-project']:
         QTimer.singleShot(100, window.new_project)
     sys.exit(app.exec())